Extracted implementation from fetch_msvc_version
